Make type of lambda int in TplDepFrame

+++ vp9/encoder/vp9_encoder.h
@@ -321,7 +321,7 @@ typedef struct TplDepFrame {
   int mi_cols;
   int base_qindex;
 #if CONFIG_NON_GREEDY_MV
-  double lambda;
+  int lambda;
   int_mv *pyramid_mv_arr[3][SQUARE_BLOCK_SIZES];
   int *mv_mode_arr[3];
   double *rd_diff_arr[3];
